Upgrade to 1.10.x and removed one level of hierarchy folder to don't need extra config for using module as it is with orchard
